Move Components::Checklist::* to Views::Controllers::Checklists::*
The three checklist components (`Contents`, `Context`, `Panel`) only
render for `ChecklistsController#show` (inline) and the
`projects/target_names#_checklist_update.erb` turbo-stream response
(same controller, refresh after target-name CUD). Single-controller
use — `app/views/controllers/checklists/` is where they belong, per
.claude/rules/phlex_conversions.md.

Folder also pluralizes: `app/components/checklist/` →
`app/views/controllers/checklists/`, matching the controller's
`controller_path = "checklists"`. The model PORO at
`app/classes/checklist.rb` keeps the singular `::Checklist` — those
are separate things.

Internal sibling references switch to unqualified names where
lexical scope allows: in `contents.rb`, `Panel.new(...)` instead of
`Components::Checklist::Panel.new(...)`; in `show.rb` (which sits in
the same `Views::Controllers::Checklists` module),
`Contents.new(...)` instead of the full path. External callers
(`ChecklistsController#checklist_context`, the turbo-stream ERB
partial) use the full `Views::Controllers::Checklists::*` path.

`render(::Components::Panel.new(...))` in panel.rb is now top-level-
qualified to disambiguate from the sibling view-side `Panel` —
different things sharing a name.
